refactor homeScreen component
