+++ src/components/Root.js
@@ -1,5 +1,8 @@
 import React from 'react'
 import { service } from '../services'
+import firebase from 'react-native-firebase'
+import AsyncStorage from '@react-native-community/async-storage'
+import { showAlert } from '../config/utils'
 
 class Root extends React.Component {
   constructor(props) {
@@ -11,7 +14,74 @@ class Root extends React.Component {
   }
 
   async componentDidMount() {
+    this.checkPermission()
+    this.createNotificationListeners()
     await service.permissionManager.requestStorage()
+  }
+
+  checkPermission() {
+    firebase.messaging().hasPermission()
+    .then(enabled => {
+      if (enabled) {
+        this.getToken();
+      } else {
+        console.log('user doesn\'t have permission')
+        this.requestPermission()
+      } 
+    });
+  }
+
+  async getToken() {
+    let fcmToken = await AsyncStorage.getItem('fcmToken');
+    if (!fcmToken) {
+      fcmToken = await firebase.messaging().getToken();
+      if (fcmToken) {
+        console.log('fcmToken', fcmToken)
+        await AsyncStorage.setItem('fcmToken', fcmToken);
+        service.apiManager.saveToken(fcmToken)
+      }
+    }
+  }
+
+  async requestPermission() {
+    try {
+      await firebase.messaging().requestPermission();
+      this.getToken();
+    } catch (error) {
+      console.log('permission rejected');
+    }
+  }
+
+  componentWillUnmount() {
+    this.notificationListener();
+    this.notificationOpenedListener();
+  }
+
+  async createNotificationListeners() {
+    /*
+    * Triggered when a particular notification has been received in foreground
+    * */
+    this.notificationListener = firebase.notifications().onNotification((notification) => {
+        const { title, body } = notification;
+        showAlert(title, body);
+    });
+  
+    /*
+    * If your app is in background, you can listen for when a notification is clicked / tapped / opened as follows:
+    * */
+    this.notificationOpenedListener = firebase.notifications().onNotificationOpened((notificationOpen) => {
+        const { title, body } = notificationOpen.notification;
+        showAlert(title, body);
+    });
+  
+    /*
+    * If your app is closed, you can check if it was opened by a notification being clicked / tapped / opened as follows:
+    * */
+    const notificationOpen = await firebase.notifications().getInitialNotification();
+    if (notificationOpen) {
+        const { title, body } = notificationOpen.notification;
+        showAlert(title, body);
+    }
   }
 
   render() {
