+++ src/services/apis/index.js
@@ -1,13 +1,11 @@
 import axios from 'axios'
 import * as config from '../../config/connectionBase'
 
-const URI = `${config.host.staging}:${config.port}`
-
 export const saveToken = async (token) => {
-  let uri = `${URI}/${config.tokens_path}`
+  let tokensPath = `${config.uri}/${config.tokens_path}`
 
   try {
-    let data = await axios.post(uri, { firebase: { token } })
+    let data = await axios.post(tokensPath, { firebase: { token } })
                     .then( resp => resp.data )
                     .catch( error => error)
     console.log(data)
@@ -17,7 +15,7 @@ export const saveToken = async (token) => {
 }
 
 export const fetch_category_children = async (category_id) => {
-  let uri = `${URI}/${config.category_path(category_id)}`
+  let uri = `${config.uri}/${config.category_path(category_id)}`
 
   try {
     let data = await axios.get(uri)
